fix node label
